+++ packages/bpk-component-calendar/src/BpkCalendarView.js
@@ -1,41 +1,19 @@
 import React, { PropTypes } from 'react';
 
 import BpkCalendarNav from './BpkCalendarNav';
-import BpkCalendarGrid, { BpkCalendarGridHeader } from './BpkCalendarGrid';
+import BpkCalendarGrid from './BpkCalendarGrid';
+import BpkCalendarGridHeader from './BpkCalendarGridHeader';
 import BpkCalendarDate from './BpkCalendarDate';
 import CustomPropTypes from './custom-proptypes';
+import addCalendarGridTransition from './BpkCalendarGridTransition';
 import './bpk-calendar.scss';
 
-const reorderWeekDays = (weekDays, weekStartsOn) => [
-  ...weekDays.slice(weekStartsOn),
-  ...weekDays.slice(0, weekStartsOn),
-];
+const TransitioningBpkCalendarGrid = addCalendarGridTransition(BpkCalendarGrid);
 
 const BpkCalendarView = (props) => {
   const classNames = ['bpk-calendar'];
 
   if (props.className) { classNames.push(props.className); }
-
-  // Sorted in [sun, mon, ..., sat]
-  const sortedWeekDays = props.daysOfWeek.slice().sort((a, b) => a.index - b.index);
-  // Ordered according to weekStartsOn, e.g. [mon, tue, ..., sun]
-  const reorderedWeekDays = reorderWeekDays(
-    sortedWeekDays,
-    props.weekStartsOn,
-  );
-
-
-  const weekend = sortedWeekDays.map(day => day.isWeekend);
-  let weekendStartIndex = -1;
-  let weekendEndIndex = -1;
-
-  if (weekend[0] && weekend[6]) { // weekend stretches over turn the of the week
-    weekendStartIndex = weekend.lastIndexOf(false) + 1;
-    weekendEndIndex = weekend.indexOf(false) - 1;
-  } else {
-    weekendStartIndex = weekend.indexOf(true);
-    weekendEndIndex = weekend.lastIndexOf(true);
-  }
 
   return (
     <div className={classNames.join(' ')}>
@@ -51,12 +29,11 @@ const BpkCalendarView = (props) => {
       />
       <BpkCalendarGridHeader
         className="bpk-calendar__grid-header"
-        weekDays={reorderedWeekDays}
+        daysOfWeek={props.daysOfWeek}
         showWeekendSeparator={props.showWeekendSeparator}
-        weekendStartIndex={weekendStartIndex}
-        weekendEndIndex={weekendEndIndex}
+        weekStartsOn={props.weekStartsOn}
       />
-      <BpkCalendarGrid
+      <TransitioningBpkCalendarGrid
         className="bpk-calendar__grid"
         DateComponent={props.DateComponent}
         dateModifiers={props.dateModifiers}
@@ -71,53 +48,51 @@ const BpkCalendarView = (props) => {
         weekStartsOn={props.weekStartsOn}
         maxDate={props.maxDate}
         minDate={props.minDate}
-
         focusedDate={props.focusedDate}
         markToday={props.markToday}
         markOutsideDays={props.markOutsideDays}
         selectedDate={props.selectedDate}
-
-        weekDays={reorderedWeekDays}
-        weekendStartIndex={weekendStartIndex}
-        weekendEndIndex={weekendEndIndex}
       />
     </div>
   );
 };
 
 BpkCalendarView.propTypes = {
-  // BpkCalendarNav & BpkCalendarGrid
-  month: PropTypes.instanceOf(Date).isRequired,
-  formatMonth: PropTypes.func.isRequired,
-
-  // BpkCalendarNav
+  // Required
   changeMonthLabel: PropTypes.string.isRequired,
+  daysOfWeek: CustomPropTypes.DaysOfWeek.isRequired,
+  formatDateFull: PropTypes.func.isRequired,
+  formatMonth: PropTypes.func.isRequired,
   id: PropTypes.string.isRequired,
   maxDate: PropTypes.instanceOf(Date).isRequired,
   minDate: PropTypes.instanceOf(Date).isRequired,
-
-  onChangeMonth: PropTypes.func,
-
-  // BpkCalendarGrid
-  DateComponent: PropTypes.func.isRequired,
-  daysOfWeek: CustomPropTypes.DaysOfWeek.isRequired,
-  formatDateFull: PropTypes.func.isRequired,
+  month: PropTypes.instanceOf(Date).isRequired,
   showWeekendSeparator: PropTypes.bool.isRequired,
   weekStartsOn: PropTypes.number.isRequired,
-
+  // Optional
   className: PropTypes.string,
+  DateComponent: PropTypes.func,
   dateModifiers: CustomPropTypes.DateModifiers,
+  focusedDate: PropTypes.instanceOf(Date),
+  markOutsideDays: PropTypes.bool,
+  markToday: PropTypes.bool,
+  onChangeMonth: PropTypes.func,
   onDateClick: PropTypes.func,
   onDateKeyDown: PropTypes.func,
   preventKeyboardFocus: PropTypes.bool,
+  selectedDate: PropTypes.instanceOf(Date),
 };
 
 BpkCalendarView.defaultProps = {
   className: null,
   DateComponent: BpkCalendarDate,
   dateModifiers: {},
-  onChangeMonth: null,
-  onDateClick: null,
+  focusedDate: null,
+  markOutsideDays: false,
+  markToday: false,
+  onChangeMonth: () => null,
+  onDateClick: () => null,
+  selectedDate: null,
   showWeekendSeparator: true,
   weekStartsOn: 1,
 };